# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_handler.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/internal_handler.APIError'
      meta:
        $ref: '#/definitions/internal_handler.Meta'
      success:
        type: boolean
    type: object
  internal_handler.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handler.bookingResponse:
    properties:
      bathhouse_id:
        type: string
      certificate_discount:
        type: integer
      comment:
        type: string
      created_at:
        type: string
      earned_points:
        type: integer
      end_time:
        type: string
      guest_count:
        type: integer
      id:
        type: string
      loyalty_discount:
        type: integer
      original_price:
        type: integer
      payment_status:
        type: string
      points_spent:
        type: integer
      promo_discount:
        type: integer
      referral_bonus_used:
        type: integer
      start_time:
        type: string
      status:
        type: string
      total_price:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  internal_handler.createBookingRequest:
    properties:
      bathhouse_id:
        type: string
      certificate_code:
        type: string
      comment:
        type: string
      end_time:
        type: string
      guest_count:
        type: integer
      promo_code:
        type: string
      start_time:
        type: string
      use_points:
        type: integer
      use_referral_bonus:
        type: integer
    type: object
...
  internal_handler.simpleMessageResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Bani bathhouse marketplace platform.
  title: Bani API
  version: "1.0"
paths:
...
  /bathhouses/{id}/bookings:
    get:
      description: Returns a paginated list of bookings for a specific bathhouse.
        Only available to bathhouse owners and representatives.
      parameters:
      - description: Bathhouse ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler.bookingResponse'
                  type: array
                meta:
                  $ref: '#/definitions/internal_handler.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: List bathhouse bookings
      tags:
      - bookings
...
  /bookings:
    get:
      description: Returns a paginated list of bookings for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler.bookingResponse'
                  type: array
                meta:
                  $ref: '#/definitions/internal_handler.Meta'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: List my bookings
      tags:
      - bookings
    post:
      consumes:
      - application/json
      description: Creates a new booking for a bathhouse. Supports loyalty points,
        referral bonus, promo codes, and gift certificates as discounts.
      parameters:
      - description: Booking data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.bookingResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Create booking
      tags:
      - bookings
  /bookings/{id}/cancel:
    patch:
      description: Cancels a booking. Clients can cancel their own bookings, owners/representatives
        can cancel bookings for their bathhouses.
      parameters:
      - description: Booking ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.simpleMessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Cancel booking
      tags:
      - bookings
  /bookings/{id}/complete:
    patch:
      description: Marks a confirmed booking as completed. Awards loyalty points.
        Only available to bathhouse owners and representatives.
      parameters:
      - description: Booking ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.bookingResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Complete booking
      tags:
      - bookings
  /bookings/{id}/confirm:
    patch:
      description: Confirms a pending booking. Only available to bathhouse owners
        and representatives.
      parameters:
      - description: Booking ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.simpleMessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Confirm booking
      tags:
      - bookings
...
  /bookings/{id}/reject:
    patch:
      description: Rejects a pending booking. Only available to bathhouse owners and
        representatives.
      parameters:
      - description: Booking ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.simpleMessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Reject booking
      tags:
      - bookings
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"